Reset bee position upon death
Because the camera is following the bee, the camera moves through the
world and also affects the background of the menu screen.  By resetting
the bee's position, the camera is moved back to the zero position.  We
might want to refactor this by moving the camera upon death or even altering
the world size.